Fix parsing of hunk headers
Summary: The present implementation of hunk headers parsing assumes that
the source and target lengths of a hunk are 0 when these are not
specified.

This is incorrect. The unidiff spec omits the length of a hunk when this
is 1, not 0: https://www.gnu.org/software/diffutils/manual/html_node/Detailed-Unified.html

This PR updates the parsing logic to default missing source and target
lengths in a hunk header to 1 and adds a test that exercises this
behaviour.

Test Plan:

Unit tests pass: `cargo test`